+++ savedsearch-core/src/main/java/oracle/sysman/emSDK/emaas/platform/savedsearch/model/SearchManager.java
@@ -1,5 +1,6 @@
 package oracle.sysman.emSDK.emaas.platform.savedsearch.model;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -43,7 +44,7 @@ public abstract class SearchManager
 	 * @param permanently
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract void deleteSearch(long searchId, boolean permanently) throws EMAnalyticsFwkException;
+	public abstract void deleteSearch(BigInteger searchId, boolean permanently) throws EMAnalyticsFwkException;
 
 	/**
 	 *
@@ -51,7 +52,7 @@ public abstract class SearchManager
 	 * @param permanently
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract void deleteTargetCard(long searchId, boolean permanently) throws EMAnalyticsFwkException;
+	public abstract void deleteTargetCard(BigInteger searchId, boolean permanently) throws EMAnalyticsFwkException;
 	/**
 	 * Resturns the search by its name dxy
 	 * @param searchName
@@ -88,7 +89,7 @@ public abstract class SearchManager
 	 * @return search
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract Search getSearch(long searchId) throws EMAnalyticsFwkException;
+	public abstract Search getSearch(BigInteger searchId) throws EMAnalyticsFwkException;
 
 	/**
 	 * @param name
@@ -96,7 +97,7 @@ public abstract class SearchManager
 	 * @return
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract Search getSearchByName(String name, long folderId) throws EMAnalyticsFwkException;
+	public abstract Search getSearchByName(String name, BigInteger folderId) throws EMAnalyticsFwkException;
 
 	/**
 	 * Returns the count of (accessible) search entities in a folder.
@@ -106,7 +107,7 @@ public abstract class SearchManager
 	 * @return count of seacrh entities
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract int getSearchCountByFolderId(long folderId) throws EMAnalyticsFwkException;
+	public abstract int getSearchCountByFolderId(BigInteger folderId) throws EMAnalyticsFwkException;
 
 	/**
 	 * Returns the list of search entities belonging to a category.
@@ -116,7 +117,7 @@ public abstract class SearchManager
 	 * @return list of search entities belonging to a category
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract List<Search> getSearchListByCategoryId(long categoryId) throws EMAnalyticsFwkException;
+	public abstract List<Search> getSearchListByCategoryId(BigInteger categoryId) throws EMAnalyticsFwkException;
 
 	/**
 	 * Returns the list of search entities contained (directly) in the specified folder.
@@ -126,7 +127,7 @@ public abstract class SearchManager
 	 * @return list of search entities (<code>null</code> if none are contained)
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract List<Search> getSearchListByFolderId(long folderId) throws EMAnalyticsFwkException;
+	public abstract List<Search> getSearchListByFolderId(BigInteger folderId) throws EMAnalyticsFwkException;
 
 	public List<Search> getSearchListByLastAccessDate(int count) throws EMAnalyticsFwkException
 	{
@@ -134,7 +135,7 @@ public abstract class SearchManager
 		return null;
 	}
 
-	public abstract List<Search> getSystemSearchListByCategoryId(long categoryId) throws EMAnalyticsFwkException;
+	public abstract List<Search> getSystemSearchListByCategoryId(BigInteger categoryId) throws EMAnalyticsFwkException;
 
 	/**
 	 * Returns the list of widgets belonging to a category.
@@ -144,7 +145,7 @@ public abstract class SearchManager
 	 * @return list of widgets belonging to a category
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract List<Search> getWidgetListByCategoryId(long categoryId) throws EMAnalyticsFwkException;
+	public abstract List<Search> getWidgetListByCategoryId(BigInteger categoryId) throws EMAnalyticsFwkException;
 
 	/**
 	 * Returns the list of widgets belonging to the categories specified by provider names
@@ -167,9 +168,9 @@ public abstract class SearchManager
 	 * @return screenshot of widget
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract ScreenshotData getWidgetScreenshotById(long widgetId) throws EMAnalyticsFwkException;
+	public abstract ScreenshotData getWidgetScreenshotById(BigInteger widgetId) throws EMAnalyticsFwkException;
 
-	public abstract Date modifyLastAccessDate(long searchId) throws EMAnalyticsFwkException;
+	public abstract Date modifyLastAccessDate(BigInteger searchId) throws EMAnalyticsFwkException;
 
 	/**
 	 * Saves a completely specified search entity in the analytics sub-system.
@@ -196,5 +197,5 @@ public abstract class SearchManager
 	 * @return
 	 * @throws EMAnalyticsFwkException
 	 */
-	public abstract String getSearchParamByName(long searchId, String paramName) throws EMAnalyticsFwkException;
+	public abstract String getSearchParamByName(BigInteger searchId, String paramName) throws EMAnalyticsFwkException;
 }
